requirements.txt
psycopg2-binary==2.8.6
pytest==6.2.3